refactor: renamed new components
