+++ src/frontend/src/eth/components/tokens/ManageTokensMenuButton.svelte
