+++ src/frontend/src/icp/components/tokens/IcManageTokensMenuButton.svelte
