+++ src/frontend/src/lib/components/tokens/TokensMenu.svelte
@@ -6,8 +6,8 @@
 	import Hr from '$lib/components/ui/Hr.svelte';
 	import { networkEthereum, networkICP } from '$lib/derived/network.derived';
 	import { erc20TokensNotInitialized } from '$eth/derived/erc20.derived';
-	import IcTokensManage from '$icp/components/tokens/IcTokensManage.svelte';
-	import TokensManage from '$eth/components/tokens/TokensManage.svelte';
+	import IcManageTokensMenuButton from '$icp/components/tokens/IcManageTokensMenuButton.svelte';
+	import ManageTokensMenuButton from '$eth/components/tokens/ManageTokensMenuButton.svelte';
 
 	let visible = false;
 	let button: HTMLButtonElement | undefined;
@@ -34,9 +34,9 @@
 			</div>
 
 			{#if $networkICP}
-				<IcTokensManage />
+				<IcManageTokensMenuButton />
 			{:else if $networkEthereum}
-				<TokensManage />
+				<ManageTokensMenuButton />
 			{/if}
 		{/if}
 	</div>
